+++ python_graph.cpp
@@ -3,18 +3,18 @@
 
 ivecvec find_optimal_partition(int n, 
 			       int T, 
-			       std::vector<float> a, 
-			       std::vector<float> b) {
+			       std::vector<double> a, 
+			       std::vector<double> b) {
   
   auto pg = PartitionGraph(n, T, a, b);
   return pg.get_optimal_subsets_extern();
   
 }
 
-float find_optimal_weight(int n,
+double find_optimal_weight(int n,
 			  int T,
-			  std::vector<float> a,
-			  std::vector<float> b) {
+			  std::vector<double> a,
+			  std::vector<double> b) {
   auto pg = PartitionGraph(n, T, a, b);
   return pg.get_optimal_weight_extern();
 
@@ -22,22 +22,22 @@ float find_optimal_weight(int n,
 
 swpair optimize_one(int n,
 		    int T,
-		    std::vector<float> a,
-		    std::vector<float> b) {
+		    std::vector<double> a,
+		    std::vector<double> b) {
 
   auto pg = PartitionGraph(n, T, a, b);
   ivecvec subsets = pg.get_optimal_subsets_extern();
-  float weight = pg.get_optimal_weight_extern();
+  double weight = pg.get_optimal_weight_extern();
 
   return std::make_pair(subsets, weight);
 }
 
 swpair sweep_best(int n,
 		  int T,
-		  std::vector<float> a,
-		  std::vector<float> b) {
+		  std::vector<double> a,
+		  std::vector<double> b) {
   
-  float best_weight = std::numeric_limits<float>::max(), weight;
+  double best_weight = std::numeric_limits<double>::max(), weight;
   ivecvec subsets;
   
   for (int i=T; i>1; --i) {
@@ -56,8 +56,8 @@ swpair sweep_best(int n,
 
 swcont sweep_parallel(int n,
 		    int T,
-		    std::vector<float> a,
-		    std::vector<float> b) {
+		    std::vector<double> a,
+		    std::vector<double> b) {
   
   ThreadsafeQueue<swpair> results_queue;
   
@@ -91,11 +91,11 @@ swcont sweep_parallel(int n,
 
 swcont sweep(int n,
 	     int T,
-	     std::vector<float> a,
-	     std::vector<float> b) {
+	     std::vector<double> a,
+	     std::vector<double> b) {
   
 
-  float weight;
+  double weight;
   ivecvec subsets;
   swcont r;
